requirements.txt
argparse==1.2.1
coverage==3.6
jsonpickle==0.4.0
nose==1.3.0
python-dateutil==2.1
six==1.3.0
wsgiref==0.1.2
